the game
it would be hilarious if i said "you lost the game", but this is no place for games. i, as the greatest programmer that has ever lived, am finally succumbing to the woke mind virus that is "version control systems". i never thought the day would come, but i believe it is finally time for me to accept that this is how the world is; however unfortunate and heart-rending it may be.