worthless
this is absolutely heart ranching. i cannot believe that i've done this. even after all of the self aggrandizing claims in my last commit, it is impossible to imagine that i spent so much time on such a small addition; i've only now realized how small the addition was while refactoring. this is cause for shame. the worst programmer to have ever programmed. truly.

- refactored main.rs into tui.rs and tui_draw.rs